backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-multipart==0.0.6
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
requests==2.31.0
python-dotenv==1.0.0

# Job aggregation system dependencies
celery==5.3.0
redis==5.0.0
flower==2.0.1
selenium==4.15.0
beautifulsoup4==4.12.0
lxml==4.9.3
dateparser==1.2.0
webdriver-manager==4.0.1
